:recycle: [fix] price 타입 Long 에서 int로 수정

+++ src/main/java/sopt/twosome/service/Favorite/FavoriteService.java
@@ -27,7 +27,7 @@ public class FavoriteService {
         Member member = memberRetriever.findMemberById(memberId);
 
         // personal이 true인 경우 할인
-        Long adjustedPrice = favoriteCreateRequest.personal()
+        int adjustedPrice = favoriteCreateRequest.personal()
                 ? favoriteCreateRequest.price() - 300
                 : favoriteCreateRequest.price();
